requirements.txt
Flask>=3.0.0
Flask-Login>=0.6.3
PyMySQL>=1.1.0
python-dotenv>=1.0.0
Werkzeug>=3.0.0
streamlit>=1.32.0
pandas>=2.0.0
